moving Hists classes from gdgem/nmx to common
