+++ prototype2/common/Hists.h
