+++ prototype2/multigrid/mgmesytec.cpp
@@ -16,7 +16,7 @@
 #include <cstring>
 #include <efu/Parser.h>
 #include <efu/Server.h>
-#include <gdgem/nmx/HistSerializer.h>
+#include <common/HistSerializer.h>
 #include <common/ReadoutSerializer.h>
 #include <iostream>
 #include <libs/include/SPSCFifo.h>
